mainly embeds and other bits and bops

+++ src/main/java/commandHandling/commands/place/queue.java
@@ -1,11 +1,18 @@
 package commandHandling.commands.place;
 
 import commandHandling.CommandContext;
+import net.dv8tion.jda.api.EmbedBuilder;
+import net.dv8tion.jda.api.entities.Message;
 import services.BotExceptions;
+import services.PermissionManager;
 import services.database.dbHandlerQ;
 
+import java.awt.*;
+import java.io.FileNotFoundException;
+import java.io.PrintStream;
 import java.util.ArrayList;
 import java.util.Random;
+import java.util.Scanner;
 import java.util.concurrent.TimeUnit;
 
 public class queue {
@@ -18,7 +25,9 @@ public class queue {
 
     private void queueing () {
         ArrayList<Integer> numbers = dbHandlerQ.getIDs();
+        ArrayList<String> commands = new ArrayList<>();
         Random random = new Random();
+        Scanner scanner;
         String file;
         int number;
 
@@ -33,21 +42,42 @@ public class queue {
         }
 
         try {
-            ctx.getMessage().getAttachments().get(0).downloadToFile("tempFiles/place/queue/" + file);
+            scanner = new Scanner(ctx.getMessage().getAttachments().get(0).retrieveInputStream().get());
         } catch (Exception e) {
             try {
-                ctx.getMessage().getReferencedMessage().getAttachments().get(0)
-                        .downloadToFile("tempFiles/place/queue/" + file);
+                scanner = new Scanner(ctx.getMessage().getReferencedMessage().getAttachments().get(0)
+                        .retrieveInputStream().get());
             } catch (Exception ee) {
                 BotExceptions.missingAttachmentException(ctx);
                 return;
             }
         }
 
-        ctx.getMessage().reply("Your file got id: " + number).queue(msg ->  {
-            msg.delete().queueAfter(32, TimeUnit.SECONDS);
-        });
-        dbHandlerQ.addToQ(number, file, ctx.getMessage().getAuthor().getId());
+        while (scanner.hasNextLine())
+            commands.add(scanner.nextLine());
+
+        if (!PermissionManager.authOwner(ctx) && commands.size() > 10000) {
+            BotExceptions.fileTooBigException(ctx);
+        } else {
+            try {
+                PrintStream printer = new PrintStream("tempFiles/place/queue/" + file);
+                for (String s : commands)
+                    printer.println(s);
+            } catch (FileNotFoundException e) {
+                e.printStackTrace();
+            }
+
+            EmbedBuilder embed = new EmbedBuilder();
+            embed.setTitle("Queue");
+            embed.setColor(new Color(0xb074ad));
+            embed.setDescription("Your file got ID " + number);
+
+            dbHandlerQ.addToQ(number, file, ctx.getMessage().getAuthor().getId());
+            ctx.getMessage().reply(embed.build()).queue(msg ->  {
+                msg.delete().queueAfter(32, TimeUnit.SECONDS);
+            });
+        }
+
         ctx.getMessage().delete().queueAfter(32, TimeUnit.SECONDS);
     }
 }
